+++ valida_generator/lib/src/validators_lib_generator.dart
@@ -72,7 +72,7 @@ ${allElements.map((e) => "import '${e.source!.uri}';").toSet().join()}
 /// found in code generation.
 class Validators with GenericValidator {
   Validators._() {
-    for (final v in <Validator>[
+    for (final v in <Validator<dynamic, dynamic>>[
       ${allElements.map((e) {
         return 'validator${_name(e)},';
       }).join()}
@@ -89,7 +89,7 @@ class Validators with GenericValidator {
 
   /// A map with all registered validators by
   /// the type of the model to validate
-  final typeMap = <Type, Validator>{};
+  final typeMap = <Type, Validator<dynamic, dynamic>>{};
 
   @override
   Validator<T, Validation<T, Object>>? validator<T>() {
@@ -105,7 +105,7 @@ class Validators with GenericValidator {
   }
 
   ${allElements.map((e) {
-        return 'static const validator${_name(e)} = Validator(${_name(e)}Validation.fromValue);';
+        return 'static const validator${_name(e)} = Validator.fromFunction(${_name(e)}Validation.fromValue);';
       }).join()}
 }
 ''';