+++ valida_generator/CHANGELOG.md
@@ -1,4 +1,10 @@
-## [0.1.0] - 2022-08-30.
+## [0.2.0] - 2023-07-08.
+
+* Use `Validator.fromFunction` for all validators
+* Require Dart 3.0.0
+* Upgrade to analyzer: ^6.0.0
+
+## [0.1.0-dev.1] - 2022-08-30.
 
 Multiple breaking changes